Break session lock out from lock surface files
