+++ src/lock-surface.c
@@ -1,86 +1,15 @@
 #include "lock-surface.h"
 
-#include "registry.h"
 #include "libwayland-shim.h"
 #include "stubbed-surface.h"
+#include "session-lock.h"
 
-#include <xdg-shell-client.h>
+#include <ext-session-lock-v1-client.h>
 #include <wayland-client-protocol.h>
 #include <assert.h>
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
-
-struct wl_display* current_display = NULL;
-struct ext_session_lock_v1* current_lock = NULL;
-static session_lock_locked_callback_t current_callback = NULL;
-static void* current_callback_data = NULL;
-static bool is_locked = false;
-
-static void session_lock_handle_locked(void* data, struct ext_session_lock_v1* session_lock) {
-    (void)data;
-    if (session_lock != current_lock) {
-        ext_session_lock_v1_unlock_and_destroy(session_lock);
-        return;
-    }
-    is_locked = true;
-    if (current_callback) {
-        current_callback(true, current_callback_data);
-    }
-}
-
-static void session_lock_handle_finished(void* data, struct ext_session_lock_v1* session_lock) {
-    (void)data;
-    if (session_lock != current_lock) {
-        ext_session_lock_v1_destroy(session_lock);
-        return;
-    }
-    is_locked = false;
-    if (current_callback) {
-        current_callback(false, current_callback_data);
-        current_callback = NULL;
-        current_callback_data = NULL;
-    }
-}
-
-static const struct ext_session_lock_v1_listener session_lock_listener = {
-    .locked = session_lock_handle_locked,
-    .finished = session_lock_handle_finished,
-};
-
-bool session_lock_lock(struct wl_display* display, session_lock_locked_callback_t callback, void* data) {
-    if (current_lock) {
-        callback(false, data);
-        return false;
-    }
-    struct ext_session_lock_manager_v1* manager = get_session_lock_global_from_display(display);
-    if (!manager) {
-        callback(false, data);
-        return false;
-    }
-    current_display = display;
-    current_lock = ext_session_lock_manager_v1_lock(manager);
-    current_callback = callback;
-    current_callback_data = data;
-    is_locked = false;
-    ext_session_lock_v1_add_listener(current_lock, &session_lock_listener, NULL);
-    return true;
-}
-
-void session_lock_unlock() {
-    if (!current_lock) return;
-    if (is_locked) {
-        ext_session_lock_v1_unlock_and_destroy(current_lock);
-        wl_display_roundtrip(current_display);
-    }
-    // if there is a current lock that is not yet locked, nulling it will cause it to be destroyed when it gets a locked
-    // or finished event
-    current_display = NULL;
-    current_lock = NULL;
-    current_callback = NULL;
-    current_callback_data = NULL;
-    is_locked = false;
-}
 
 static void lock_surface_handle_configure(
     void* data,
@@ -136,18 +65,14 @@ struct lock_surface_t lock_surface_make(struct wl_output* output) {
     return ret;
 }
 
-void lock_surface_map(struct lock_surface_t* self, void* lock_callback_data) {
+void lock_surface_map(struct lock_surface_t* self) {
     if (self->lock_surface) {
         return;
     }
 
+    struct ext_session_lock_v1* current_lock = session_lock_get_active_lock();
     if (!current_lock) {
         fprintf(stderr, "failed to create lock surface, no current lock in place\n");
-        return;
-    }
-
-    if (lock_callback_data != current_callback_data) {
-        fprintf(stderr, "failed to create lock surface, callback data doesn't match\n");
         return;
     }
 
@@ -193,7 +118,7 @@ static bool xdg_wm_base_get_xdg_surface_hook(
     if (self) {
         *ret_proxy = xdg_surface_server_get_xdg_surface(&self->super, (struct xdg_wm_base*)proxy, wl_surface);
         return true;
-    } else if (current_lock) {
+    } else if (session_lock_get_active_lock()) {
         // A new XDG surface is being created while the screen is locked, but it's not a lock surface. Few possibilities
         // 1. it's going to be a toplevel
         // 2. it's going to be a popup on a non-lock toplevel
